fix: first load logs req and labels
- fix issue of trying to load logs when there is no query
- fix issue with adding / loading labels after remove or add mannualy
- refactor of loadLogs service into split files
- fix adding / removing labels from log line or table after pipe symbol
- remove unnecesary loading events
