+++ src/actions/helpers/index.js
@@ -3,3 +3,4 @@ export * from "./prepareCols";
 export * from "./prepareVectorRows";
 export * from "./setColumnsData";
 export * from "./setColumnsValue";
+
