+++ src/components/LabelBrowser/helpers/onQueryValid.js
@@ -1,3 +1,10 @@
 export default function onQueryValid(query) {
-    return query !== "{" && query !== "}" && query !== "{}" && query !== ""; // TODO: make a proper query validation
+    return (
+        query &&
+        query !== "{" &&
+        query !== "}" &&
+        query !== "{}" &&
+        query !== "" &&
+        query?.length >= 7
+    ); // TODO: make a proper query validation
 }
