+++ src/components/StatusBar/components/apiselector/ApiSelector.js
@@ -13,15 +13,15 @@ import {
     ApiSelectorStyled,
 } from "../../styled";
 import loadLabels from "../../../../actions/loadLabels";
+import onQueryValid from "../../../LabelBrowser/helpers/onQueryValid";
 
 export function ApiSelector() {
-    const apiUrl = useSelector((store) => store.apiUrl);
-    const apiError = useSelector((store) => store.apiErrors);
+    const { apiUrl, apiError, query } = useSelector((store) => store);
     const [editedUrl, setEditedUrl] = useState(apiUrl);
     const [apiSelectorOpen, setApiSelectorOpen] = useState(false);
     const dispatch = useDispatch();
     const API_URL = "API URL";
-    
+
     useEffect(() => {
         setEditedUrl(apiUrl);
     }, []);
@@ -32,13 +32,16 @@ export function ApiSelector() {
     }, [apiUrl]);
 
     useEffect(() => {
-        if (apiError.length > 0) {
+        if (apiError?.length > 0) {
             setApiSelectorOpen(true);
             dispatch(setLogs([]));
             dispatch(setMatrixData([]));
         } else {
             setApiSelectorOpen(false);
-            dispatch(loadLogs());
+            if (query?.length > 0 && onQueryValid(query)) {
+                 dispatch(loadLogs());
+            }
+           
         }
     }, [apiError]);
